requirements.txt
airtest==1.3.0.1
cached-property==1.5.2
certifi==2023.5.7
charset-normalizer==3.2.0
colored==1.4.4
decorator==5.1.1
Deprecated==1.2.14
deprecation==2.1.0
facebook-wda==1.4.6
ffmpeg-python==0.2.0
filelock==3.12.2
future==0.18.3
hrpc==1.0.9
idna==3.4
Jinja2==3.1.2
loguru==0.7.0
logzero==1.7.0
MarkupSafe==2.1.3
mss==6.1.0
numpy==1.21.6
opencv-contrib-python==4.6.0.66
packaging==23.1
Pillow==9.5.0
pocoui==1.0.89
py==1.11.0
python-xlib==0.33
pywinauto==0.6.3
requests==2.31.0
retry==0.9.2
simple-tornado==0.2.2
simplejson==3.19.1
six==1.16.0
tabulate==0.9.0
tidevice==0.11.0
tornado==6.2
urllib3==1.26.16
websocket-client==0.48.0
wrapt==1.15.0
